Put the survey function in the survey module
